Tried to implement functions in class but it makes no sense. I need to think of something to build in a smart way. Now I go back to clean pyproj.
